requirements.txt
Flask==2.3.2
flask-cors==3.0.10
pymongo==4.6.0
protobuf==4.23.3            # required by prophet
cmdstanpy==1.0.8            # backend for Prophet
prophet==1.1.4
holidays==0.33
pandas==2.0.3
numpy==1.26.4
python-dotenv==1.0.0
gunicorn==23.0.0         # For production server on Render